Install PyYAML package
YAML is actually a good fit for the new configuration format: The new
configuration format might require some duplication for common use-
cases which can be avoided with YAML references. This keeps the
configuration and code simple while making good use of YAML.

Plus, YAML is a superset of JSON, so configuration can be generated
with any configuration generator that can output JSON (like Nix).